requirements.txt
openai>=1.30.0
aiosqlite>=0.20.0
edge-tts>=6.1.12
httpx>=0.27.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
loguru>=0.7.2
rich>=13.7.0
